# Conflicts:
#	pom.xml
#	src/main/java/se/sundsvall/supportmanagement/service/CommunicationService.java
#	src/main/resources/application-it.yml
#	src/main/resources/application-junit.yml
#	src/main/resources/application.yml
#	src/test/java/se/sundsvall/supportmanagement/service/CommunicationServiceTest.java
#	src/test/resources/api/openapi.yaml

+++ src/main/java/se/sundsvall/supportmanagement/api/EmailIntegrationConfigResource.java
@@ -24,8 +24,11 @@ import se.sundsvall.dept44.common.validators.annotation.ValidMunicipalityId;
 import se.sundsvall.dept44.problem.Problem;
 import se.sundsvall.dept44.problem.violations.ConstraintViolationProblem;
 import se.sundsvall.supportmanagement.api.model.config.EmailIntegration;
+import se.sundsvall.supportmanagement.integration.db.model.enums.ProtectedResource;
+import se.sundsvall.supportmanagement.service.AccessControlService;
 import se.sundsvall.supportmanagement.service.config.EmailIntegrationConfigService;
 
+import static generated.se.sundsvall.accessmapper.Access.AccessLevelEnum.RW;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.HttpHeaders.LOCATION;
 import static org.springframework.http.MediaType.ALL_VALUE;
@@ -45,9 +48,11 @@ import static se.sundsvall.supportmanagement.Constants.NAMESPACE_VALIDATION_MESS
 class EmailIntegrationConfigResource {
 
 	private final EmailIntegrationConfigService service;
+	private final AccessControlService accessControlService;
 
-	EmailIntegrationConfigResource(final EmailIntegrationConfigService service) {
+	EmailIntegrationConfigResource(final EmailIntegrationConfigService service, final AccessControlService accessControlService) {
 		this.service = service;
+		this.accessControlService = accessControlService;
 	}
 
 	@PostMapping(consumes = APPLICATION_JSON_VALUE, produces = ALL_VALUE)
@@ -62,6 +67,8 @@ class EmailIntegrationConfigResource {
 		@Parameter(name = "namespace", description = "Namespace", example = "MY_NAMESPACE") @Pattern(regexp = NAMESPACE_REGEXP, message = NAMESPACE_VALIDATION_MESSAGE) @PathVariable final String namespace,
 		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
 		@Valid @NotNull @RequestBody final EmailIntegration emailIntegration) {
+
+		accessControlService.verifyNamespaceAuthorization(namespace, municipalityId, ProtectedResource.EMAIL_INTEGRATION_CONFIG, RW);
 
 		service.create(emailIntegration, namespace, municipalityId);
 
@@ -100,6 +107,8 @@ class EmailIntegrationConfigResource {
 		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
 		@Valid @NotNull @RequestBody final EmailIntegration emailIntegration) {
 
+		accessControlService.verifyNamespaceAuthorization(namespace, municipalityId, ProtectedResource.EMAIL_INTEGRATION_CONFIG, RW);
+
 		service.replace(emailIntegration, namespace, municipalityId);
 		return noContent()
 			.header(CONTENT_TYPE, ALL_VALUE)
@@ -118,6 +127,8 @@ class EmailIntegrationConfigResource {
 	ResponseEntity<Void> deleteEmailIntegrationConfig(
 		@Parameter(name = "namespace", description = "Namespace", example = "MY_NAMESPACE") @Pattern(regexp = NAMESPACE_REGEXP, message = NAMESPACE_VALIDATION_MESSAGE) @PathVariable final String namespace,
 		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId) {
+
+		accessControlService.verifyNamespaceAuthorization(namespace, municipalityId, ProtectedResource.EMAIL_INTEGRATION_CONFIG, RW);
 
 		service.delete(namespace, municipalityId);
 		return noContent()
